audio-mcp/pyproject.toml
[project]
name = "audio-mcp"
version = "0.1.0"
description = "Audio Processing MCP Server"
authors = [
    {name = "MCP Tutorial", email = "tutorial@example.com"}
]
dependencies = [
    "fastmcp>=0.1.0",
    "pydub>=0.25.1",
    "mutagen>=1.47.0",
    "asyncio-subprocess>=0.1.0"
]

[project.optional-dependencies]
ai = [
    "openai-whisper>=20231117",
    "torch>=2.0.0",
    "torchaudio>=2.0.0"
]
tts = [
    "gTTS>=2.4.0",
    "pyttsx3>=2.90"
]
all = [
    "audio-mcp[ai,tts]"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
audio-mcp = "audio_server:main"